[rename] hooks -> service 폴더명 변경
