+++ src/components/Board/CommentList.jsx
@@ -6,8 +6,8 @@ import PropTypes from 'prop-types';
 import styled from 'styled-components';
 import BoardComment from './BoardComment';
 import Typing from './Typing';
-import { UserContext } from '../../hooks/UserContext';
-import { BoardContext } from '../../hooks/BoardContext';
+import { UserContext } from '../../service/UserContext';
+import { BoardContext } from '../../service/BoardContext';
 import theme from '../../styles/theme';
 
 const TypingContainer = styled.div`
