+++ src/components/Board/Typing.jsx
@@ -3,7 +3,7 @@ import styled from 'styled-components';
 import axios from 'axios';
 import PropTypes from 'prop-types';
 import registerComment from '../../assets/images/ic_send.svg';
-import { replaceNewLines } from '../../hooks/Utils';
+import { replaceNewLines } from '../../service/Utils';
 import theme from '../../styles/theme';
 
 const InputWrapper = styled.div`
