+++ src/pages/NoticeDetail.jsx
@@ -2,8 +2,8 @@ import React, { useState, useEffect, useContext } from 'react';
 import styled from 'styled-components';
 import axios from 'axios';
 import { useNavigate, useParams } from 'react-router-dom';
-import UserAPI from '../hooks/UserAPI';
-import { UserContext } from '../hooks/UserContext';
+import UserAPI from '../service/UserAPI';
+import { UserContext } from '../service/UserContext';
 import BoardHeader from '../components/Board/NoticeHeader';
 import AttachButton from '../components/Board/AttachButton';
 import CommentList from '../components/Board/CommentList';
